gcp/requirements.txt
tensorflow-cpu==2.7
google-cloud-storage==1.16.1
Pillow==8.4.0
numpy>=1.22,<1.24
protobuf==3.20.*